refactor: Auth 컴포넌트 및 로직 분리, svgr 설정, layout padding 설정, 불필요한 페이지 archive 처리
