+++ src/app/auth/hooks/useLoginForm.ts
@@ -1,0 +1,55 @@
+import { useState } from "react";
+
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+
+import { loginSchema, LoginFormData } from "@/validation/auth-validation";
+
+export function useLoginForm() {
+  const [isLoading, setIsLoading] = useState(false);
+  const [serverError, setServerError] = useState<string | null>(null);
+
+  const form = useForm<LoginFormData>({
+    resolver: zodResolver(loginSchema),
+    defaultValues: {
+      email: "",
+      password: "",
+    },
+    mode: "onSubmit",
+  });
+
+  const onSubmit = async (data: LoginFormData) => {
+    setIsLoading(true);
+    setServerError(null);
+
+    try {
+      // API 예시
+      await new Promise((resolve) => setTimeout(resolve, 500));
+      console.log("로그인 시도:", data);
+
+      // TODO: 실제 로그인 API 연동
+      // const response = await fetch('/api/auth/login', {
+      //   method: 'POST',
+      //   headers: { 'Content-Type': 'application/json' },
+      //   body: JSON.stringify(data),
+      // });
+
+      // if (!response.ok) {
+      //   throw new Error('로그인에 실패했습니다. 이메일과 비밀번호를 확인해주세요.');
+      // }
+
+      // window.location.href = '/dashboard';
+    } catch (error) {
+      console.error("로그인 실패:", error);
+      setServerError(
+        error instanceof Error
+          ? error.message
+          : "로그인 중 오류가 발생했습니다.",
+      );
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  return { form, isLoading, serverError, onSubmit };
+}
