+++ src/app/layout.tsx
@@ -21,8 +21,8 @@ const pretendard = localFont({
 // });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Chulcheck",
+  description: "출석체크 서비스",
 };
 
 export default function RootLayout({
@@ -32,7 +32,11 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="ko">
-      <body className={`${pretendard.variable} antialiased`}>{children}</body>
+      <body className={`${pretendard.variable} antialiased`}>
+        <div className="min-h-screen">
+          <div className="px-4 py-6 sm:px-6 lg:px-8">{children}</div>
+        </div>
+      </body>
     </html>
   );
 }
