+++ src/components/svg/logo_wordmark.tsx
@@ -1,0 +1,5 @@
+import Logo from "@/assets/logo_wordmark.svg";
+
+export default function LogoWordmark() {
+  return <Logo />;
+}
