+++ src/components/ui/form.tsx
@@ -15,7 +15,7 @@ import {
 } from "react-hook-form";
 
 import { Label } from "@/components/ui/label";
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/utils";
 
 const Form = FormProvider;
