+++ src/components/ui/label.tsx
@@ -4,7 +4,7 @@ import * as React from "react";
 
 import * as LabelPrimitive from "@radix-ui/react-label";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/utils/utils";
 
 function Label({
   className,
