+++ src/lib/type/svg.d.ts
@@ -1,0 +1,5 @@
+declare module "*.svg" {
+  import React from "react";
+  const svg: React.FC<React.SVGProps<SVGSVGElement>>;
+  export default svg;
+}
